TIP-860: Move ProductNormalizer for Versioning from Tool to Pim/Enrichment
