+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/serializers_versionning.yml
@@ -1,6 +1,6 @@
 parameters:
     pim_serializer.normalizer.flat.category.class: Akeneo\Pim\Enrichment\Component\Category\Normalizer\Versioning\CategoryNormalizer
-
+    pim_serializer.normalizer.flat.product.class: Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\ProductNormalizer
 
 services:
     pim_versioning.serializer.normalizer.flat.category:
@@ -8,5 +8,10 @@ services:
         arguments:
             - '@pim_catalog.normalizer.standard.category'
             - '@pim_versioning.serializer.normalizer.flat.label_translation'
+        tags:
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    pim_versioning.serializer.normalizer.flat.product:
+        class: '%pim_serializer.normalizer.flat.product.class%'
         tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
